# lab2/data-service 2/target/generated-sources/openapi-feign/api/openapi.yaml
openapi: 3.0.0
info:
  description: API for managing student data
  title: Data Service API
  version: 1.0.0
servers:
- description: Local server
  url: http://localhost:8080
paths:
  /api/students:
    get:
      operationId: getAllStudents
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/Student"
                type: array
          description: List of students
      summary: Get all students
      tags:
      - Students
      x-accepts:
      - application/json
    post:
      operationId: createStudent
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Student"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
          description: Student created
      summary: Create a new student
      tags:
      - Students
      x-content-type: application/json
      x-accepts:
      - application/json
  /api/students/{id}:
    delete:
      operationId: deleteStudent
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "204":
          description: Student deleted
      summary: Delete student
      tags:
      - Students
      x-accepts:
      - application/json
    get:
      operationId: getStudentById
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
          description: Student found
        "404":
          description: Student not found
      summary: Get student by ID
      tags:
      - Students
      x-accepts:
      - application/json
    put:
      operationId: updateStudent
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Student"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
          description: Student updated
      summary: Update student
      tags:
      - Students
      x-content-type: application/json
      x-accepts:
      - application/json
components:
  schemas:
    Student:
      example:
        firstName: firstName
        lastName: lastName
        id: 0
        email: email
      properties:
        id:
          format: int64
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        email:
          format: email
          type: string
      required:
      - email
      - firstName
      - lastName
      type: object