# lab2/data-service 2/target/generated-sources/openapi/src/main/java/ru/kondrya/data_service/api/StudentsApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.15.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package ru.kondrya.data_service.api;

import ru.kondrya.data_service.model.Student;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.lang.Nullable;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-12-15T19:24:59.053757190Z[Etc/UTC]", comments = "Generator version: 7.15.0")
@Validated
@Tag(name = "Students", description = "the Students API")
public interface StudentsApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    public static final String PATH_CREATE_STUDENT = "/api/students";
    /**
     * POST /api/students : Create a new student
     *
     * @param student  (required)
     * @return Student created (status code 201)
     */
    @Operation(
        operationId = "createStudent",
        summary = "Create a new student",
        tags = { "Students" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Student created", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Student.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = StudentsApi.PATH_CREATE_STUDENT,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<Student> createStudent(
        @Parameter(name = "Student", description = "", required = true) @Valid @RequestBody Student student
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"firstName\" : \"firstName\", \"lastName\" : \"lastName\", \"id\" : 0, \"email\" : \"email\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    public static final String PATH_DELETE_STUDENT = "/api/students/{id}";
    /**
     * DELETE /api/students/{id} : Delete student
     *
     * @param id  (required)
     * @return Student deleted (status code 204)
     */
    @Operation(
        operationId = "deleteStudent",
        summary = "Delete student",
        tags = { "Students" },
        responses = {
            @ApiResponse(responseCode = "204", description = "Student deleted")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = StudentsApi.PATH_DELETE_STUDENT
    )
    
    default ResponseEntity<Void> deleteStudent(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    public static final String PATH_GET_ALL_STUDENTS = "/api/students";
    /**
     * GET /api/students : Get all students
     *
     * @return List of students (status code 200)
     */
    @Operation(
        operationId = "getAllStudents",
        summary = "Get all students",
        tags = { "Students" },
        responses = {
            @ApiResponse(responseCode = "200", description = "List of students", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = Student.class)))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = StudentsApi.PATH_GET_ALL_STUDENTS,
        produces = { "application/json" }
    )
    
    default ResponseEntity<List<Student>> getAllStudents(
        
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ { \"firstName\" : \"firstName\", \"lastName\" : \"lastName\", \"id\" : 0, \"email\" : \"email\" }, { \"firstName\" : \"firstName\", \"lastName\" : \"lastName\", \"id\" : 0, \"email\" : \"email\" } ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    public static final String PATH_GET_STUDENT_BY_ID = "/api/students/{id}";
    /**
     * GET /api/students/{id} : Get student by ID
     *
     * @param id  (required)
     * @return Student found (status code 200)
     *         or Student not found (status code 404)
     */
    @Operation(
        operationId = "getStudentById",
        summary = "Get student by ID",
        tags = { "Students" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Student found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Student.class))
            }),
            @ApiResponse(responseCode = "404", description = "Student not found")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = StudentsApi.PATH_GET_STUDENT_BY_ID,
        produces = { "application/json" }
    )
    
    default ResponseEntity<Student> getStudentById(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"firstName\" : \"firstName\", \"lastName\" : \"lastName\", \"id\" : 0, \"email\" : \"email\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    public static final String PATH_UPDATE_STUDENT = "/api/students/{id}";
    /**
     * PUT /api/students/{id} : Update student
     *
     * @param id  (required)
     * @param student  (required)
     * @return Student updated (status code 200)
     */
    @Operation(
        operationId = "updateStudent",
        summary = "Update student",
        tags = { "Students" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Student updated", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Student.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = StudentsApi.PATH_UPDATE_STUDENT,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<Student> updateStudent(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id,
        @Parameter(name = "Student", description = "", required = true) @Valid @RequestBody Student student
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"firstName\" : \"firstName\", \"lastName\" : \"lastName\", \"id\" : 0, \"email\" : \"email\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}